sara-ai-assistant/pyproject.toml
[project]
name = "sara-ai-assistant"
version = "1.0.0"
description = "Smart Autonomous Responsive Assistant by Nikhil Badal"
authors = [{name = "Nikhil Badal"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.2.0",
    "livekit-agents>=0.8.0",
    "livekit-plugins-openai>=0.6.0",
    "livekit-plugins-deepgram>=0.6.0",
    "livekit-plugins-google>=0.6.0",
    "google-generativeai>=0.8.0",
    "openai>=1.0.0",
    "groq>=0.4.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
    "flask-socketio>=5.3.0",
    "psutil>=5.9.0",
    "pyautogui>=0.9.54",
    "pycaw>=20240210",
    "comtypes>=1.2.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "speech-recognition>=3.10.0",
    "pyttsx3>=2.90",
    "wikipedia>=1.4.0",
    "wolframalpha>=5.0.0",
    "spotipy>=2.23.0",
    "youtube-dl>=2021.12.17",
    "pywin32>=306"
]

[project.scripts]
sara = "server:main"
sara_voice = "agent_sara:dev"
sara_ui = "ui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0"
]